Fix formatting

+++ src/java.base/share/classes/java/util/LinkedList.java
@@ -310,7 +310,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public E removeFirst(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {        final Node<E> f = first;
+    public E removeFirst(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {
+        final Node<E> f = first;
         if (f == null)
             throw new NoSuchElementException();
         return unlinkFirst(f);
@@ -324,7 +325,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public E removeLast(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {        final Node<E> l = last;
+    public E removeLast(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {
+        final Node<E> l = last;
         if (l == null)
             throw new NoSuchElementException();
         return unlinkLast(l);
@@ -337,7 +339,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public void addFirst(@Growable @GuardSatisfied LinkedList<E> this, E e) {        linkFirst(e);
+    public void addFirst(@Growable @GuardSatisfied LinkedList<E> this, E e) {
+        linkFirst(e);
     }
 
     /**
@@ -349,7 +352,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public void addLast(@Growable @GuardSatisfied LinkedList<E> this, E e) {        linkLast(e);
+    public void addLast(@Growable @GuardSatisfied LinkedList<E> this, E e) {
+        linkLast(e);
     }
 
     /**
@@ -389,7 +393,8 @@ public class LinkedList<E>
     @EnsuresNonEmpty("this")
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean add(@Growable @GuardSatisfied LinkedList<E> this, E e) {        linkLast(e);
+    public boolean add(@Growable @GuardSatisfied LinkedList<E> this, E e) {
+        linkLast(e);
         return true;
     }
 
@@ -409,7 +414,8 @@ public class LinkedList<E>
     @ReleasesNoLocks
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean remove(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @GuardSatisfied @Nullable @UnknownSignedness Object o) {        if (o == null) {
+    public boolean remove(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @GuardSatisfied @Nullable @UnknownSignedness Object o) {
+        if (o == null) {
             for (Node<E> x = first; x != null; x = x.next) {
                 if (x.item == null) {
                     unlink(x);
@@ -441,7 +447,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean addAll(@Growable @GuardSatisfied LinkedList<E> this, Collection<? extends E> c) {        return addAll(size, c);
+    public boolean addAll(@Growable @GuardSatisfied LinkedList<E> this, Collection<? extends E> c) {
+        return addAll(size, c);
     }
 
     /**
@@ -461,7 +468,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean addAll(@Growable @GuardSatisfied LinkedList<E> this, @NonNegative int index, Collection<? extends E> c) {        checkPositionIndex(index);
+    public boolean addAll(@Growable @GuardSatisfied LinkedList<E> this, @NonNegative int index, Collection<? extends E> c) {
+        checkPositionIndex(index);
 
         Object[] a = c.toArray();
         int numNew = a.length;
@@ -505,7 +513,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public void clear(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {        // Clearing all of the links between nodes is "unnecessary", but:
+    public void clear(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {
+        // Clearing all of the links between nodes is "unnecessary", but:
         // - helps a generational GC if the discarded nodes inhabit
         //   more than one generation
         // - is sure to free memory even if there is a reachable Iterator
@@ -548,7 +557,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public E set(@Replaceable @GuardSatisfied LinkedList<E> this, @NonNegative int index, E element) {        checkElementIndex(index);
+    public E set(@Replaceable @GuardSatisfied LinkedList<E> this, @NonNegative int index, E element) {
+        checkElementIndex(index);
         Node<E> x = node(index);
         E oldVal = x.item;
         x.item = element;
@@ -566,7 +576,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public void add(@Growable @GuardSatisfied LinkedList<E> this, @NonNegative int index, E element) {        checkPositionIndex(index);
+    public void add(@Growable @GuardSatisfied LinkedList<E> this, @NonNegative int index, E element) {
+        checkPositionIndex(index);
 
         if (index == size)
             linkLast(element);
@@ -585,7 +596,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public E remove(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @NonNegative int index) {        checkElementIndex(index);
+    public E remove(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @NonNegative int index) {
+        checkElementIndex(index);
         return unlink(node(index));
     }
 
@@ -740,7 +752,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public @Nullable E poll(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {        final Node<E> f = first;
+    public @Nullable E poll(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {
+        final Node<E> f = first;
         return (f == null) ? null : unlinkFirst(f);
     }
 
@@ -753,7 +766,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public E remove(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {        return removeFirst();
+    public E remove(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {
+        return removeFirst();
     }
 
     /**
@@ -765,7 +779,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean offer(@Growable LinkedList<E> this, E e) {        return add(e);
+    public boolean offer(@Growable LinkedList<E> this, E e) {
+        return add(e);
     }
 
     // Deque operations
@@ -778,7 +793,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean offerFirst(@Growable LinkedList<E> this, E e) {        addFirst(e);
+    public boolean offerFirst(@Growable LinkedList<E> this, E e) {
+        addFirst(e);
         return true;
     }
 
@@ -791,7 +807,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean offerLast(@Growable LinkedList<E> this, E e) {        addLast(e);
+    public boolean offerLast(@Growable LinkedList<E> this, E e) {
+        addLast(e);
         return true;
     }
 
@@ -833,7 +850,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public @Nullable E pollFirst(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {        final Node<E> f = first;
+    public @Nullable E pollFirst(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {
+        final Node<E> f = first;
         return (f == null) ? null : unlinkFirst(f);
     }
 
@@ -847,7 +865,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public @Nullable E pollLast(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {        final Node<E> l = last;
+    public @Nullable E pollLast(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this) {
+        final Node<E> l = last;
         return (l == null) ? null : unlinkLast(l);
     }
 
@@ -862,7 +881,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public void push(@Growable @GuardSatisfied LinkedList<E> this, E e) {        addFirst(e);
+    public void push(@Growable @GuardSatisfied LinkedList<E> this, E e) {
+        addFirst(e);
     }
 
     /**
@@ -878,7 +898,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public E pop(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {        return removeFirst();
+    public E pop(@Shrinkable @GuardSatisfied @NonEmpty @CanShrink LinkedList<E> this) {
+        return removeFirst();
     }
 
     /**
@@ -892,7 +913,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean removeFirstOccurrence(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @GuardSatisfied @Nullable @UnknownSignedness Object o) {        return remove(o);
+    public boolean removeFirstOccurrence(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @GuardSatisfied @Nullable @UnknownSignedness Object o) {
+        return remove(o);
     }
 
     /**
@@ -906,7 +928,8 @@ public class LinkedList<E>
      */
     // @SideEffectsOnly("this")
     @DoesNotUnrefineReceiver("modifiability")
-    public boolean removeLastOccurrence(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @GuardSatisfied @Nullable @UnknownSignedness Object o) {        if (o == null) {
+    public boolean removeLastOccurrence(@Shrinkable @GuardSatisfied @CanShrink LinkedList<E> this, @GuardSatisfied @Nullable @UnknownSignedness Object o) {
+        if (o == null) {
             for (Node<E> x = last; x != null; x = x.prev) {
                 if (x.item == null) {
                     unlink(x);
